+++ NINA/Utility/AvalonDock/PaneTemplateSelector.cs
@@ -27,6 +27,7 @@ using NINA.ViewModel.Equipment.FilterWheel;
 using NINA.ViewModel.Equipment.Focuser;
 using NINA.ViewModel.Equipment.Guider;
 using NINA.ViewModel.Equipment.Rotator;
+using NINA.ViewModel.Equipment.Switch;
 using NINA.ViewModel.Equipment.Telescope;
 using System.Windows;
 using System.Windows.Controls;
@@ -72,6 +73,8 @@ namespace NINA.Utility.AvalonDock {
         public DataTemplate ThumbnailTemplate { get; set; }
 
         public DataTemplate FocusTargetsTemplate { get; set; }
+
+        public DataTemplate SwitchTemplate { get; set; }
 
         public override System.Windows.DataTemplate SelectTemplate(object item, System.Windows.DependencyObject container) {
             var itemAsLayoutContent = item as LayoutContent;
@@ -142,6 +145,10 @@ namespace NINA.Utility.AvalonDock {
 
             if (item is FocusTargetsVM) {
                 return FocusTargetsTemplate;
+            }
+
+            if (item is SwitchVM) {
+                return SwitchTemplate;
             }
 
             return base.SelectTemplate(item, container);